Fix integration tests build
svn path=/trunk/matlab/; revision=2880


Former-commit-id: 61fce305d13554a290a373a85d81037c0feb04e0

+++ maintenance/TeamCity/TeamCity_runtests.m
@@ -182,20 +182,20 @@ try %#ok<TRYNC>
             delete(fullfile(targetdir,'mxdom2defaulthtml.xsl'));
         end
 
-        if OPT.PublishCoverage
-            mtr.MTestPublisher.publishcoverage(mtr.ProfileInfo,'TargetDir',fullfile(targetdir,'coverage'));
-    	    delete('OetTestCoverage.zip');
-            zip('OetTestCoverage',{fullfile(targetdir,'coverage','*.*')});
-        end
-
-        %% zip result and remove target dir
-        if OPT.Publish
-            mtr.MTestPublisher.publishtestsoverview(mtr,'TargetDir',fullfile(targetdir,'testoverview'));
-            delete('OetTestResult.zip');
-            zip('OetTestResult',{fullfile(targetdir,'testoverview','*.*')});
-        end
-
-        if OPT.Publish || OPT.PublishCoverage
+        if any(~[mtr.Tests.Igrnore]) && (OPT.Publish || OPT.PublishCoverage)
+            if OPT.PublishCoverage
+                mtr.MTestPublisher.publishcoverage(mtr.ProfileInfo,'TargetDir',fullfile(targetdir,'coverage'));
+                delete('OetTestCoverage.zip');
+                zip('OetTestCoverage',{fullfile(targetdir,'coverage','*.*')});
+            end
+            
+            %% zip result and remove target dir
+            if OPT.Publish
+                mtr.MTestPublisher.publishtestsoverview(mtr,'TargetDir',fullfile(targetdir,'testoverview'));
+                delete('OetTestResult.zip');
+                zip('OetTestResult',{fullfile(targetdir,'testoverview','*.*')});
+            end
+            
             rmdir(targetdir,'s');
         end
